Make consent requirement UI more flexible.
This change makes it so consent requirements can be a little more
flexible in their display. The majority of the change is plumbing that
allows a requirement to override the default text for the actions (Allow
or Block). This also makes it so that if the requirement doesn't specify
a data use description, we don't show that block.

Testing done:
Used in conjunction with a specialized requirement for asking users to
accept a privacy policy and terms of service.

Reviewed at https://reviews.reviewboard.org/r/9957/