Add a consent requirement for policies (privacy and terms of service).
This change adds a common consent requirement for policy enforcement. In
the case where an admin has specified their privacy policy and/or terms
of service, they likely are running a public service where they'll need
their users to accept them. This consent requirement isn't actually used
inside of Djblets, but can be added to a registry in a Djblets-using
application and checked therein.

Testing done:
Used in conjunction with a change in Review Board.

Reviewed at https://reviews.reviewboard.org/r/9959/