Fix a crash when generating PII-safe URLs with non-strings in kwargs.
When generating PII-safe URLs, we attempt to analyze the content of any
values coming into the view from the keyword arguments. If something
comes in that's not a string, this can crash, and that'll happen any
time extra context is provided to a view (like a form class).

We now check the value type prior to checking the content.

Testing Done:
Unit tests pass.

Tested with a view that crashed. Verified this fixed it.

Reviewed at https://reviews.reviewboard.org/r/9963/