Add an extension hook for registering consent requirements.
This introduces `djblets.privacy.consent.hooks.ConsentRequirementHook`,
which registers a `ConsentRequirement`, managing its registration as the
extension initializes or shuts down. This is a very simple hook that
just makes use of `BaseRegistryHook`, feeding in the registry
dynamically on use (to avoid creating a registry unnecessarily or
prematurely).

Testing Done:
Unit tests passed.

Used this in an extension successfully.

Reviewed at https://reviews.reviewboard.org/r/9964/