Fix loading datagrids due to recent template render compatibility changes.
The recent changes to switch to the compatibility functions for
loading/rendering templates broke datagrids, due to an argument that was
ultimately left out of the implementation of `render()`. The datagrid
was trying to render without using a `RequestContext`, since it already
created a slimmed-down context containing necessary request information
in order to avoid performance penalties.

This updates the call to explicitly use `render_to_string()` instead, so
that we can avoid `RequestContext`, and returns an `HttpResponse` for
that rendered content.

Testing Done:
Verified that the dashboard loads again.

Reviewed at https://reviews.reviewboard.org/r/10272/