Prevent an unwanted save when enabling an extension.
When enabling an extension, we'd mark the `RegisteredExtension` entry as
installed and then save it. This was only really necessary the first
time an extension was installed, but we were still doing it every time.

This change simply makes the update conditional, and reduces the expense
of the save by specifying that only the `installed` field should be
written to the database.

Testing Done:
Unit tests passed.

Manually tested enabling and disabling extensions. Verified that
extensions were always loaded when they should be and would reload when
next initialized.

Reviewed at https://reviews.reviewboard.org/r/10379/