Improve styling, defaults, and customizability for config forms.
The config forms styles in Djblets were somewhat incomplete, leaving
some important formatting (such as field errors and help text) to the
consuming applications. It also just didn't look very good by default in
some areas, such as the text fields, and required too much internal
knowledge in order to adjust the available widths of the forms.

This change backports a lot of the styling from RBCommons, which is a
heavier user of the config forms:

* Fixes the styling for error displays and radio button lists
* Fixes up lots of spacing and sizing issues for buttons, labels,
  paragraphs, and other parts of the UI
* Adds a clean, rounded border for text fields, and fleshes out the
  types of text fields that are styled
* Lightens the backgrounds, so the contrast is better

It also adds a mixin function for setting the widths of the page, so
a consumer can increase/decrease the width of the navigation pane and
forms areas.

Testing Done:
Tested all the config forms pages in Review Board and RBCommons, looking
for regressions. Compared to the previous styling.

Reviewed at https://reviews.reviewboard.org/r/10405/