Modernize the structure, naming, and rules for config forms CSS styles.
The config forms CSS styles were hobbled together from a bunch of styles
found in Review Board, RBCommons, and Splat, and in the process of
moving these into Djblets, we ended up with some unwanted rules and some
very specific styling and naming that didn't make sense to keep in
Djblets. We also have some new styles coming for config forms that was
hard to put in without a bit of a cleanup.

This change moves the styles to our new CSS naming standards, helping
create a more logical, organized, documented set of components we can
build upon, and in the process helps us move away from a bunch of
unwanted styles. We still have those styles, and the old names, for
backwards-compatibility purposes, but going forward we have a stronger
foundation to build upon, and a clear path to deprecation.

Some of these styles might come back in some other form later, but the
intent is to go back to something closer to a bare minimum. That said,
none of this is set in stone yet, and there are many form-related styles
that deserve its own cleanup (and are not considered "stable" at this
point).

Testing Done:
Tested that all the backwards-compatibility parts work with the My Account
page in Review Board and both the My Account and Team Admin pages in
RBCommons.

Made use of some of the new styles in some upcoming work.

Reviewed at https://reviews.reviewboard.org/r/10423/