Add a standard config forms container class for consistent margins.
In the old config forms, spacing between elements and within the parent
were partly defined by the stylesheet and partly left up to the
consumer. This meant that some uses had to be specially tweaked to get
elements to look correct. Some of the built-in styles made this more
difficult, with unwanted borders and margins in places.

This change defines a new `.djblets-l-config-forms-container` class,
which is intended to be used whenever there's a form element, text
description, set of buttons, or anything else that's displayed in a
config form. This defines a standard 1em margin between other containers
and between the parent (which is now recommended to provide no padding).

It also allows the contents within to be recessed into the parent
container, replacing the old `.box-recessed`. This is useful for
components like lists.

Testing Done:
Made use of this with some new UI backed by config forms.

Tested that all existing config forms UIs we have were unaffected by the
style changes, meaning the backwards compatibility worked.

Reviewed at https://reviews.reviewboard.org/r/10428/