Add modifiers to make config form containers flush against sides.
The config forms containers had a rule to remove the top/bottom margins
for the first/last `.djblets-l-config-forms-container` element within a
parent. This worked for the usage in RBCommons, where we wanted some
components (like lists) to appear toward the top/bottom of a box without
adding additional margins. This was done with `:first-child` and
`:last-child` rules, which didn't work unless the layout was just
right.

This change removes those rules and replaces them with new modifiers for
making the container flush against the top or bottom, which just removes
any margins/borders. This is easier to reason about and tweak in
consuming applications.

Testing Done:
Tested this with some container usage for integrations, and played with
styles in the browser.

Reviewed at https://reviews.reviewboard.org/r/10441/