Add support for deleting integration configs from the list.
This redoes the management of the integration configuration list to make
use of the dynamic Config Forms list components. This allows us to
define a Delete action, which will delete the configuration and
automatically update the list. This is much easier than going into the
configuration page (assuming the configuration page offers a Delete
button).

This necessitated some changes to the loading of the page. Now, instead
of building the list up-front, we serialize the data for the UI, and
it's passed in through the Config Forms list machinery. While that's
loading, a "Loading integrations" spinner will be displayed, so the list
doesn't look empty.

Further changes will build on this to better construct the popup display.

Testing Done:
Tested this in both the admin UI and in the RBCommons Team Admin UI.

Unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/10439/