Fix the name of the generated djblets-integrations JavaScript file.
Due to a copy/paste error, the resulting JavaScript file for the
`djblets-integrations` JavaScript bundle was named `extensions.min.js`.
This resulted in the file not being generated and used in production, as
the original `extensions.min.js` was taking precedence.

This is a trivial change that fixes the name of the file.