Use the fallback avatar service if the default service is None.
The introduction of the fallback avatar service was intended to always
provide a useful service if the user had opted out of any other
configured services, but it didn't catch the case where the user didn't
have an explicit service configured and the administrator had configured
the default to be no service.

This change covers this case, only including the default as a viable
option if it's not `None`, and always including the fallback service.

Testing Done:
Unit tests passed.

Verified manually that we no longer get errors with a `None` avatar
service if the default is `None`, and that instead we get the fallback
service.

Bugs closed: 4800

Reviewed at https://reviews.reviewboard.org/r/10516/