Normalize the text and HTML body types for EmailMessage.
`EmailMessage` now accepts both byte strings and Unicode strings for the
text and HTML bodies. These are normalized to Unicode strings, which are
compatible with the implementations on both Python 2 and 3, and across
the range of Django versions.

Testing Done:
Djblets and Review Board unit tests pass on Python 2/3 with Django 1.6
and 1.11.

Reviewed at https://reviews.reviewboard.org/r/10471/