Fix sorting of integrations on Python 3.x.
The code to sort integrations was using the entire dictionary as a key.
This *happened* to work on Python 2.x, but failed entirely on 3.x. We
now use the integration name as the key instead, which is what was
intended originally.

Testing Done:
Manually tested on Python 2.7 and 3.7.

Unit tests pass on both.

Reviewed at https://reviews.reviewboard.org/r/10524/