Fix a small regression with Pygments and Markdown in unit tests.
The unit tests for the Djblets Markdown extensions were assuming a
setting would be turned off by default, which wasn't the case. This
resulted in some markup in the rendered HTML that we didn't want.

This is a trivial change that adds the configuration flag to turn off
this setting before rendering in the unit tests.