Add evolutions for updating app labels.
We have a few apps that were given more explicit, namespaced app labels
in Djblets 2.0.dev. Since we're supporting Django 1.11 and Django
Evolution 2.0 now, we need to be able to clearly distinguish between app
labels and to make sure that the evolution project signature is using
the right app labels for the right apps.

To help this along, this change adds evolutions that define the app
label for some of our apps, taking the guesswork out of it and ensuring
that the project signature is evolved correctly.

Testing Done:
Unit tests passed.

Tested evolving the database on Django 1.6 and 1.11 with Review Board 4.0.

Reviewed at https://reviews.reviewboard.org/r/10603/