Rename Settings to ExtensionSettings and improve documentation.
The object that managed an extension's settings was called `Settings`,
which was a bit too generic and capable of conflicts. It also didn't
match the naming scheme of most other core Extension classes.

This change renames it to `ExtensionSettings` and improves its
documentation. It keeps an older `Settings` alias around for any legacy
purposes, which we'll remove in Djblets 3.0.

Testing Done:
Unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/10641/