Add a namespace for datagrid variables.
This updates the Djblets datagrid CSS to add a
`#djblets-ns-ui.datagrids()` namespace for all the various colors,
padding, and other styles we use throughout the stylesheet.

The new namespace is useful as a reference to other components looking
to fit in with the presentation of the datagrids, or ones that are
looking to alter the presentation in a way that avoids hard-coding
duplicate values.

The style of the datagrid in Djblets remains as it was. While it's dated
and should probably be stripped of much of its default presentation,
that would need to be tackled separately.

Testing Done:
Tested that datagrids still looked and worked correctly.

Reviewed at https://reviews.reviewboard.org/r/10662/