Fix clicking on bare rows in datagrids.
The datagrids once had a facility for the entire row to act as a link to
the given object. Over time as we made changes, this functionality had
been about halfway erased, making it so only cells with `.has-link`
worked as links despite the entire row having a pointer icon. This
change fixes up things so that the URL is set correctly on the `tr`
element's data, and recreates a javascript handler for clicks that are
otherwise unhandled.

Testing Done:
Clicked on various cells within different dashboards and checked to
ensure that the correct page was loaded. For cells that contained their
own link (for example, users), saw that the row-level link was ignored
in favor of the cell-level link. Verified that clicks on
interaction-type cells (stars, checkboxes) did not activate the link.

Reviewed at https://reviews.reviewboard.org/r/10712/